QGDS-34 breadcrumbs JS logic updated and filenames standardised
